Arrange benchmarks to avoid using unnecessary parameters
